ccp module added

+++ MCAL_layer/mcal_std_types.h
@@ -16,12 +16,12 @@
 
 
 /* Section :Data Type Declarations */
-typedef  unsigned char      uint8 ;
-typedef  unsigned short     uint16 ;
-typedef  unsigned int       uint32 ;
-typedef  signed char        sint8 ;
-typedef  signed short       sint16 ;
-typedef  signed int         sint32 ;
+typedef  unsigned char       uint8 ;
+typedef  unsigned short      uint16 ;
+typedef  unsigned long       uint32 ;
+typedef  signed char         sint8 ;
+typedef  signed short        sint16 ;
+typedef  signed long         sint32 ;
 
 typedef  uint8              Std_ReturnType ;
